requirements/base.txt
Django>=1.11.3
django-admin-bootstrapped>=2.5.7
djangorestframework>=3.6.3
